restructure before options changes
adds boost_program_options requirement, future feature
