+++ CMakeLists.txt
@@ -51,7 +51,10 @@ option(WARMUP "WARMUP" OFF)
 set(VERSION_MAJOR 2)
 set(VERSION_MINOR 2)
 set(VERSION_PATCH 2)
-
+# -------------------------------------------------------------------------
+# -------------------------------------------------------------------------
+find_package(Boost REQUIRED COMPONENTS program_options)
+include_directories(${Boost_INCLUDE_DIRS})
 # -------------------------------------------------------------------------
 # Capture GIT SHA in include/dromajo_sha.h
 # -------------------------------------------------------------------------
@@ -171,6 +174,7 @@ add_library(dromajo_cosim STATIC
         src/riscv_machine.cpp
         src/softfp.cpp
         src/term_io.cpp
+        src/uart.cpp
         src/virtio.cpp
         src/LiveCache.cpp
         )
@@ -178,7 +182,9 @@ add_library(dromajo_cosim STATIC
 add_executable(dromajo src/dromajo.cpp)
 add_executable(dromajo_cosim_test src/dromajo_cosim_test.cpp)
 
-include_directories(include external ${CMAKE_CURRENT_BINARY_DIR})
+target_link_libraries(dromajo Boost::program_options)
+target_link_libraries(dromajo_cosim_test Boost::program_options)
+target_link_libraries(dromajo_cosim Boost::program_options)
 
 if (GOLDMEM)
   target_link_libraries(dromajo dromajo_cosim gold)
@@ -187,6 +193,8 @@ else ()
   target_link_libraries(dromajo dromajo_cosim)
   target_link_libraries(dromajo_cosim_test dromajo_cosim)
 endif ()
+
+include_directories(include external ${CMAKE_CURRENT_BINARY_DIR})
 
 # -------------------------------------------------------------------
 # Change the name of the copied file if SIMPOINT option is enabled
